requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
matplotlib>=3.3.0
pillow>=8.0.0
tqdm>=4.50.0
tensorboard>=2.4.0
scipy>=1.5.0
pandas>=1.1.0